refactor: FeatureInterface -> FeatureTypes, FeatureProcessInterface -> FeatureProcessTypes
