+++ module/JepRiaShowcase/App/client-react/src/features/feature-process/pages/List.tsx
@@ -3,7 +3,7 @@ import { useDispatch } from "react-redux";
 import { useHistory, useLocation, useParams } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { Grid } from "@jfront/ui-core";
-import { FeatureProcess } from "../api/FeatureProcessInterface";
+import { FeatureProcess } from "../api/FeatureProcessTypes";
 import { findFeatureProcess } from "../api/FeatureProcessApi";
 import { setState, Workstates } from "../../../app/WorkstateSlice";
 import { setCurrentFeatureProcess } from "../featureProcessSlice";
