+++ module/JepRiaShowcase/App/client-react/src/features/feature/pages/Search.tsx
@@ -9,8 +9,8 @@ import { DatePicker } from "@jfront/ui-core";
 import { CheckBoxGroup } from "@jfront/ui-core";
 import { CheckBox } from "@jfront/ui-core";
 import { TextInput } from "@jfront/ui-core";
-import { FeatureSearchTemplate } from "../api/FeatureInterface";
-import { FeatureStatusOptions } from "../../feature-process/api/FeatureProcessInterface";
+import { FeatureSearchTemplate } from "../api/FeatureTypes";
+import { FeatureStatusOptions } from "../../feature-process/api/FeatureProcessTypes";
 import { getFeatureStatusOptions } from "../../feature-process/api/FeatureProcessApi";
 import { setState, Workstates } from "../../../app/WorkstateSlice";
 import { selectSearchSubmit, selectSearchTemplate, submitSearch } from "../featureSearchSlice";
